// @flywave/flywave-mbstyle-datasource/src/MBTileDataEmitter.d.ts
import { DecodedTile } from '@flywave/flywave-datasource-protocol';
import { TileKey } from '@flywave/flywave-geoutils';
import * as THREE from 'three';
import { EvaluatedLayer } from './MBLayerEvaluator';
import { ILineGeometry, IPolygonGeometry } from '@flywave/flywave-vectortile-datasource/IGeometryProcessor';
import { DecodeInfo } from '@flywave/flywave-vectortile-datasource/DecodeInfo';
export declare class MBTileDataEmitter {
    private m_tileKey;
    private m_decodeInfo;
    private m_zoom;
    private m_geometries;
    private m_techniqueIndex;
    private m_techniques;
    private m_layerToTechniqueIndex;
    private m_glyphLookup;
    constructor(m_tileKey: TileKey, m_decodeInfo: DecodeInfo, m_zoom: number);
    setGlyphLookup(lookup: {
        getMetrics: (font: string, char: string) => any;
    }): void;
    setBearing(bearing: number): void;
    private m_bearing;
    private static s_spriteInfos;
    static setSpriteInfos(infos: Map<string, {
        width: number;
        height: number;
        pixelRatio?: number;
    }> | null): void;
    private m_textGeometries;
    private m_textPathGeometries;
    private m_poiGeometries;
    private m_stringCatalog;
    private m_stringIndex;
    private m_heatmapPoints;
    addHeatmapPoint(pos: THREE.Vector3, weight: number, radius: number, techniqueIdx: number, radiusExpr?: any, properties?: Record<string, any>): void;
    private static exprDependsOnZoom;
    private getStringIndex;
    private m_extents;
    setExtents(extents: number): void;
    get extents(): number;
    private getOrCreateGeometry;
    private project;
    private projectWorld;
    private m_currentZOffset;
    private m_maxGeometryHeight;
    private noteGeometryHeight;
    private static scaleColorByAlpha;
    private static deriveAutoBorderColor;
    private resolveZOffset;
    private extractSortKey;
    private paintToTechniqueProps;
    private static evaluatedCacheKey;
    private getOrCreateTechniqueIndex;
    processFillFeature(layerName: string, extents: number, geometry: IPolygonGeometry[], properties: Record<string, any>, featureId: string | number | undefined, matchedLayers: EvaluatedLayer[]): void;
    private emitFillOutline;
    private emitExtrudedPolygon;
    private m_lineInterleaved;
    private m_lineIndices;
    private m_lineGroupStarts;
    private m_lineSortKeys;
    private m_lineAttr;
    private m_preExtrudedLines;
    processLineFeature(layerName: string, extents: number, geometry: ILineGeometry[], properties: Record<string, any>, featureId: string | number | undefined, matchedLayers: EvaluatedLayer[]): void;
    private splitPathByAngle;
    private offsetPolyline;
    private emitRibbonBorder;
    private getOrCreateBorderTechniqueIndex;
    private emitRibbonFill;
    private emitRibbonBody;
    private emitRibbonCaps;
    private evaluateFloorLineWidth;
    private dashWorldFor;
    private getOrCreateRibbonTechniqueIndex;
    private getOrCreateOutlineTechniqueIndex;
    private getLineGeometries;
    processPointFeature(layerName: string, extents: number, points: THREE.Vector3[], properties: Record<string, any>, featureId: string | number | undefined, matchedLayers: EvaluatedLayer[]): void;
    private emitTextGeometry;
    private emitPoiGeometry;
    getDecodedTile(): DecodedTile;
}
//# sourceMappingURL=MBTileDataEmitter.d.ts.map